requirements.txt
aniso8601==9.0.1
apturl==0.5.2
async-generator==1.10
attrs==23.1.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.3
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
chardet==4.0.0
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cs50==9.2.5
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
exceptiongroup==1.1.1
fake-useragent==1.1.3
fasteners==0.14.1
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
future==0.18.2
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==1.1.2
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyboard==0.13.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
powerline-gitstatus==1.3.1
powerline-status==2.8.2
protobuf==4.23.3
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
sqlparse==0.4.4
systemd-python==234
termcolor==2.3.0
trio==0.22.0
trio-websocket==0.10.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==2.3.4
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
